backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1
httpx==0.28.1
gspread==6.1.4
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.159.0
